refactor: move processEventBatch to shared
